+++ database/migrations/2016_10_05_000026_create_test_result_table.php
@@ -17,7 +17,7 @@ class CreateTestResultTable extends Migration
             $table->integer('test_id')->unsigned();
             $table->integer('user_id')->unsigned();
             $table->integer('mark_type_id')->unsigned()->nullable()->default(NULL);
-            $table->tinyInteger('attempt')->nullable()->default(NULL);
+            $table->smallInteger('attempt')->nullable()->default(NULL);
             $table->dateTime('date_time');
 
             $table->foreign('user_id')->references('id')->on('user')->onDelete('cascade');
